Use new rrq offload configuration API.
rrq 0.7.20 has changed a bit how the offload gets configured. Instead of
being a global property that must be configured ahead of time, the path
to the offload directory is now set per-controller/worker. The offload
threshold size is similar, although for workers it is set through the
worker configuration.